Split code into modules
- Create models.py for data classes
- Move helper functions to utils.py
- Keep async logic in parse.py
