+++ app/parse.py
@@ -1,41 +1,28 @@
 import asyncio
-import csv
 import logging
 import sys
 import time
-from dataclasses import astuple, dataclass, fields
-from urllib.parse import urljoin
 
 import aiohttp
-from bs4 import BeautifulSoup, Tag
+from bs4 import BeautifulSoup
 
+from app.models import Author, Quote
+from app.utils import (
+    BASE_URL,
+    get_author_url,
+    get_next_page_url,
+    parse_author_bio,
+    parse_single_quote,
+    write_authors_to_csv,
+    write_quotes_to_csv,
+)
 
-BASE_URL = "https://quotes.toscrape.com/"
 MAX_CONCURRENT_REQUESTS = 3
-
-
-@dataclass
-class Author:
-    name: str
-    born_date: str
-    born_location: str
-    description: str
-
-
-@dataclass
-class Quote:
-    text: str
-    author: str
-    tags: list[str]
-
-
-QUOTE_FIELDS = [field.name for field in fields(Quote)]
-AUTHOR_FIELDS = [field.name for field in fields(Author)]
 
 
 logging.basicConfig(
     level=logging.INFO,
-    format="[%(levelname)8s]: %(message)s",
+    format="[%(levelname)s]: %(message)s",
     handlers=[
         logging.FileHandler("parser.log", encoding="utf-8"),
         logging.StreamHandler(sys.stdout),
@@ -43,52 +30,11 @@ logging.basicConfig(
 )
 
 
-def parse_single_quote(quote_div: Tag) -> Quote:
-    """Parse a single quote from the quote div element."""
-    return Quote(
-        text=quote_div.select_one(".text").text,
-        author=quote_div.select_one(".author").text,
-        tags=[tag.text for tag in quote_div.select(".tag")],
-    )
-
-
-def get_single_page_quotes(page_soup: Tag) -> list[Quote]:
-    """Extract all quotes from a single page."""
-    quotes_divs = page_soup.select(".quote")
-    return [parse_single_quote(quote_div) for quote_div in quotes_divs]
-
-
-def get_next_page_url(page_soup: Tag) -> str | None:
-    """Get URL of the next page if it exists."""
-    next_button = page_soup.select_one("li.next a")
-
-    if next_button:
-        return urljoin(BASE_URL, next_button["href"])
-
-    return None
-
-
-def get_author_url(quote_div: Tag) -> str:
-    """Get author page URL from the quote div element."""
-    about_link = quote_div.select_one(".author + a")
-    return urljoin(BASE_URL, about_link["href"])
-
-
 async def get_page_content(session: aiohttp.ClientSession, url: str) -> str:
     """Get HTML content of the page."""
     async with session.get(url) as response:
         response.raise_for_status()
         return await response.text()
-
-
-def parse_author_bio(page_soup: Tag) -> Author:
-    """Parse author biography from the author's page."""
-    born_date = page_soup.select_one(".author-born-date").text
-    born_location = page_soup.select_one(".author-born-location").text
-    description = page_soup.select_one(".author-description").text.strip()
-    name = page_soup.select_one(".author-title").text
-
-    return Author(name, born_date, born_location, description)
 
 
 async def process_single_page(
@@ -154,7 +100,7 @@ async def get_page_urls(
             if not next_link:
                 break
 
-            current_url = urljoin(BASE_URL, next_link["href"])
+            current_url = get_next_page_url(soup)
 
         return urls
 
@@ -195,28 +141,8 @@ async def get_all_quotes() -> tuple[list[Quote], list[Author]]:
     return all_quotes, all_authors
 
 
-def write_quotes_to_csv(quotes: list[Quote], output_path: str) -> None:
-    """Write quotes to a CSV file."""
-    with open(output_path, "w", newline="", encoding="utf-8") as file:
-        writer = csv.writer(file)
-        writer.writerow(QUOTE_FIELDS)
-        writer.writerows([astuple(quote) for quote in quotes])
-
-    logging.info(f"Quotes have been saved to {output_path}")
-
-
-def write_authors_to_csv(authors: list[Author], output_path: str) -> None:
-    """Write authors to a CSV file."""
-    with open(output_path, "w", newline="", encoding="utf-8") as file:
-        writer = csv.writer(file)
-        writer.writerow(AUTHOR_FIELDS)
-        writer.writerows([astuple(author) for author in authors])
-
-    logging.info(f"Authors have been saved to {output_path}")
-
-
 async def async_main(quotes_path: str, authors_path: str) -> None:
-    """Scrape all quotes and authors, save them to a CSV files."""
+    """Scrape all quotes and authors, save them to CSV files."""
     logging.info("Starting quotes scraping...")
     quotes, authors = await get_all_quotes()
     write_quotes_to_csv(quotes, quotes_path)
